requirement.txt
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
psycopg2==2.9.6
whitenoise==6.4.0
